Extract out common functionality of Select, ComboBox into a common ListBox component
